Fix PlayerState aetheryte id field types

+++ FFXIVClientStructs/FFXIV/Client/Game/UI/PlayerState.cs
@@ -51,10 +51,10 @@ public unsafe partial struct PlayerState {
     [FieldOffset(0x2B1)] public byte GCRankMaelstrom;
     [FieldOffset(0x2B2)] public byte GCRankTwinAdders;
     [FieldOffset(0x2B3)] public byte GCRankImmortalFlames;
-    [FieldOffset(0x2B4)] public byte HomeAetheryteId; // FIXME: ushort
+    [FieldOffset(0x2B4)] public ushort HomeAetheryteId;
     [FieldOffset(0x2B6)] public byte FavouriteAetheryteCount;
-    [FieldOffset(0x2B8)] public fixed byte FavouriteAetheryteArray[4]; // FIXME: ushort[4]
-    [FieldOffset(0x2C0)] public byte FreeAetheryteId; // FIXME: ushort
+    [FieldOffset(0x2B8)] public fixed ushort FavouriteAetheryteArray[4];
+    [FieldOffset(0x2C0)] public ushort FreeAetheryteId;
     [FieldOffset(0x2C2)] public ushort FreeAetherytePlayStationPlus;
     [FieldOffset(0x2C4)] public uint BaseRestedExperience;
